refactor(form-builder): hoist checkerboard component
