+++ packages/@sanity/form-builder/src/components/Checkerboard.tsx
